remove getWorldAbi_deprecated, update tests
